refactor(link): extract common functionalities to separate class
