+++ src/Dropdown/index.ts
@@ -1,3 +1,3 @@
 export { SgdsDropdown } from './sgds-dropdown'
 export { SgdsDropdownItem } from './sgds-dropdown-item'
-export type { DropdownButtonVariant, DropDirection} from './dropdown'
+export type { DropdownButtonVariant, DropDirection} from '../utils/dropdown-element'
