+++ src/Dropdown/sgds-dropdown.ts
@@ -2,7 +2,7 @@ import { html } from "lit";
 import { customElement } from "lit/decorators.js";
 import {  ref } from "lit/directives/ref.js";
 import styles from "./sgds-dropdown.scss";
-import { DropdownElement } from "./dropdown";
+import { DropdownElement } from "../utils/dropdown-element";
 
 export type DropDirection = "left" | "right" | "up" | "down";
 @customElement("sgds-dropdown")
