+++ src/Mainnav/sgds-mainnav-item.scss
@@ -5,7 +5,7 @@
   --mainnav-item-theme-color: #5925DC;
 }
 
-.nav-item {
+li {
   height: 100%;
 }
