+++ src/Sidenav/sgds-sidenav-link.ts
@@ -1,25 +1,10 @@
 import { html } from "lit";
-import SgdsElement from "../utils/sgds-element";
-import { customElement, property } from "lit/decorators.js";
+import LinkElement from "../utils/link-element";
+import { customElement } from "lit/decorators.js";
 import styles from "./sgds-sidenav-link.scss";
+import { classMap } from "lit/directives/class-map.js";
 
 @customElement("sgds-sidenav-link")
-export class SgdsSidenavLink extends SgdsElement {
+export class SgdsSidenavLink extends LinkElement {
   static styles = styles;
-
-  @property({ type: Boolean })
-  active = false;
-
-  @property({ type: String })
-  href = "";
-
-  render() {
-    return html`
-      <li>
-        <a href="${this.href}" class="nav-link ${this.active ? "active" : null}"
-          ><slot></slot
-        ></a>
-      </li>
-    `;
-  }
 }
