+++ src/utils/dropdown-element.ts
@@ -3,10 +3,10 @@ import { Dropdown } from "bootstrap";
 import * as Popper from "@popperjs/core";
 import type { StrictModifiers } from "@popperjs/core";
 import { createRef, Ref } from "lit/directives/ref.js";
-import mergeDeep from "../utils/mergeDeep";
-import genId from "../utils/generateId";
-import { SgdsDropdownItem } from "./sgds-dropdown-item";
-import SgdsElement from "../utils/sgds-element";
+import mergeDeep from "./mergeDeep";
+import genId from "./generateId";
+import { SgdsDropdownItem } from "../Dropdown/sgds-dropdown-item";
+import SgdsElement from "./sgds-element";
 
 const ARROW_DOWN = "ArrowDown";
 const ARROW_UP = "ArrowUp";
